+++ CHANGELOG.md
@@ -8,10 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- graph::Builder.refspecs(), to control graph walk.
+
 ### Changed
 
-- (BREAKING) GitGraph::new, add argument refspecs to control graph walk.
-
+- (BREAKING) GitGraph::new, optional arguments are moved to builder functions.
 
 ### Removed